+++ app/controllers/menu_items_controller.rb
@@ -1,40 +1,50 @@
 class MenuItemsController < ApplicationController
   def new
-    @category_id = params[:id]
-    render "new"
+    if current_user.owner?
+      @category_id = params[:id]
+      render "new"
+    end
   end
 
   def create
-    item = MenuItem.new(
-      menu_category_id: params[:menu_category_id],
-      name: params[:name],
-      description: params[:description],
-      price: params[:price],
-    )
-    item.save()
-    redirect_to "/menu_admin"
+    if current_user.owner?
+      item = MenuItem.new(
+        menu_category_id: params[:menu_category_id],
+        name: params[:name],
+        description: params[:description],
+        price: params[:price],
+      )
+      item.save()
+      redirect_to "/menu_admin"
+    end
   end
 
   def edit
-    id = params[:id]
-    @item = MenuItem.find(id)
+    if current_user.owner?
+      id = params[:id]
+      @item = MenuItem.find(id)
+    end
   end
 
   def update
-    id = params[:id]
-    item = MenuItem.find(id)
-    item.update(
-      name: params[:name],
-      description: params[:description],
-      price: params[:price],
-    )
-    redirect_to "/menu_admin"
+    if current_user.owner?
+      id = params[:id]
+      item = MenuItem.find(id)
+      item.update(
+        name: params[:name],
+        description: params[:description],
+        price: params[:price],
+      )
+      redirect_to "/menu_admin"
+    end
   end
 
   def destroy
-    id = params[:id]
-    item = MenuItem.find(id)
-    item.destroy
-    redirect_to request.referrer
+    if current_user.owner?
+      id = params[:id]
+      item = MenuItem.find(id)
+      item.destroy
+      redirect_to request.referrer
+    end
   end
 end
